Add checking time expiration
