+++ data.php
@@ -1,0 +1,40 @@
+<?php
+$projects = ['Входящие', 'Учеба', 'Работа', 'Домашние дела', 'Авто'];
+$tasks = [
+    [
+        'name' => 'Собеседование в IT компании',
+        'date' => '24.04.2019',
+        'category' => 'Работа',
+        'done' => true
+    ],
+    [
+        'name' => 'Выполнить тестовое задание',
+        'date' => '	25.12.2019',
+        'category' => 'Работа',
+        'done' => false
+    ],
+    [
+        'name' => 'Сделать задание первого раздела',
+        'date' => '21.12.2019',
+        'category' => 'Учеба',
+        'done' => true
+    ],
+    [
+        'name' => 'Встреча с другом',
+        'date' => '	24.04.2019',
+        'category' => 'Входящие',
+        'done' => false
+    ],
+    [
+        'name' => 'Купить корм для кота',
+        'date' => null,
+        'category' => 'Домашние дела',
+        'done' => false
+    ],
+    [
+        'name' => 'Заказать пиццу',
+        'date' => null,
+        'category' => 'Домашние дела',
+        'done' => false
+    ]
+];
